Added check if the df is ok
Check if there is a Expense.csv file at all, and if there is non a creats one.
Checks if all needed columns are in the dataframe.
Check for every row if all entries are in a format that can be used. If there is an entry in a format that can not be used returns a list with the rows in which to find the faulty entries and says in which format the entries should be in.